# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me/settings": {
      "get": {
        "operationId": "Viewer_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Get viewer settings",
        "tags": [
          "viewer"
        ]
      },
      "patch": {
        "operationId": "Viewer_updateSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Update viewer settings",
        "tags": [
          "viewer"
        ]
      }
    },
...
  },
  "info": {
    "title": "Matcha API",
    "description": "Canonical API contract for the Matcha backend. Target OpenAPI 3.1 where the active toolchain remains smooth.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "matcha-bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Bearer access token returned by Matcha authentication endpoints."
      }
    },
...
  }
}